+++ src/software/ai/hl/stp/tactic/BUILD
@@ -143,49 +143,6 @@ cc_test(
 )
 
 cc_library(
-    name = "intercept_ball_tactic",
-    srcs = ["intercept_ball_tactic.cpp"],
-    hdrs = [
-        "intercept_ball_fsm.h",
-        "intercept_ball_tactic.h",
-    ],
-    deps = [
-        "//software/ai/hl/stp/tactic/move:move_tactic",
-        ":tactic",
-        "//shared:constants",
-        "//software/ai/hl/stp/action:intercept_ball_action",
-        "//software/ai/hl/stp/action:stop_action",
-        "//software/ai/intent:move_intent",
-        "//software/logger",
-    ],
-)
-
-cc_test(
-    name = "intercept_ball_fsm_test",
-    srcs = ["intercept_ball_fsm_test.cpp"],
-    deps = [
-        ":intercept_ball_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_test(
-    name = "simulated_intercept_ball_tactic_test",
-    srcs = ["simulated_intercept_ball_tactic_test.cpp"],
-    deps = [
-        "//software/ai/hl/stp/tactic:intercept_ball_tactic",
-        "//software/simulated_tests:simulated_tactic_test_fixture",
-        "//software/simulated_tests:simulated_test_main",
-        "//software/simulated_tests/validation:validation_function",
-        "//software/simulated_tests/validation_functions",
-        "//software/test_util",
-        "//software/time:duration",
-        "//software/world",
-    ],
-)
-
-cc_library(
     name = "passer_tactic",
     srcs = ["passer_tactic.cpp"],
     hdrs = ["passer_tactic.h"],
@@ -351,7 +308,6 @@ cc_library(
         ":crease_defender_tactic",
         ":defense_shadow_enemy_tactic",
         ":goalie_tactic",
-        ":intercept_ball_tactic",
         ":kickoff_chip_tactic",
         ":passer_tactic",
         ":penalty_kick_tactic",
@@ -363,6 +319,7 @@ cc_library(
         ":tactic",
         "//software/ai/hl/stp/tactic/chip:chip_tactic",
         "//software/ai/hl/stp/tactic/get_behind_ball:get_behind_ball_tactic",
+        "//software/ai/hl/stp/tactic/intercept_ball:intercept_ball_tactic",
         "//software/ai/hl/stp/tactic/kick:kick_tactic",
         "//software/ai/hl/stp/tactic/move:move_tactic",
         "//software/ai/hl/stp/tactic/stop:stop_tactic",
