+++ src/software/ai/hl/stp/tactic/intercept_ball/intercept_ball_fsm_test.cpp
@@ -1,4 +1,4 @@
-#include "software/ai/hl/stp/tactic/intercept_ball_fsm.h"
+#include "software/ai/hl/stp/tactic/intercept_ball/intercept_ball_fsm.h"
 
 #include <gtest/gtest.h>
